requirements.txt
# HotGigs.ai Backend Requirements
# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# Database and ORM
supabase==1.0.4
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Authentication and Security
bcrypt==4.0.1
cryptography==41.0.4
PyJWT==2.8.0

# API and HTTP
requests==2.31.0
urllib3==2.0.4

# Data Processing and Analysis
pandas==2.1.1
numpy==1.25.2
openpyxl==3.1.2

# AI and Machine Learning
openai==0.28.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
transformers==4.33.2
torch==2.0.1

# Document Processing and OCR
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
Pillow==10.0.1
PyPDF2==3.0.1
python-docx==0.8.11

# Google Drive Integration
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.3

# Email Processing
email-validator==2.0.0
imapclient==2.3.1

# Task Queue and Background Jobs
celery==5.3.2
redis==4.6.0

# Validation and Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Performance and Monitoring
psutil==5.9.5
memory-profiler==0.61.0

# File Handling
zipfile36==0.1.3
pathlib==1.0.1

# Async Support
gevent==23.7.0
eventlet==0.33.3

# Image Processing
imageio==2.31.3
matplotlib==3.7.2

# Text Processing
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# Workflow and Automation
schedule==1.2.0
APScheduler==3.10.4

# Logging and Monitoring
structlog==23.1.0
sentry-sdk==1.32.0

# Configuration Management
pydantic==2.3.0
pydantic-settings==2.0.3

# HTTP Client
httpx==0.24.1

# JSON Processing
orjson==3.9.7

# Environment and System
python-magic==0.4.27
click==8.1.7

# Additional AI Libraries
langchain==0.0.292
tiktoken==0.5.1
faiss-cpu==1.7.4

# Document Analysis
textstat==0.7.3
readability==0.3.1

# Fraud Detection
imagehash==4.3.1
exifread==3.0.0

# Performance Optimization
cachetools==5.3.1
lru-dict==1.2.0

# Data Validation
cerberus==1.3.4
jsonschema==4.19.0

# Workflow Engine
prefect==2.13.1

# Additional Utilities
tqdm==4.66.1
rich==13.5.2
typer==0.9.0

